Вьюхи с апи перенесены в апи (логично), гитигнор, добро пожаловать
