+++ .gitignore
@@ -1,0 +1,23 @@
+# Python cache
+__pycache__/
+*.pyc
+*.pyo
+*.pyd
+
+# Virtual environment
+.venv/
+
+# Django stuff:
+*.log
+*.pot
+*.pyc
+db.sqlite3
+media/
+
+# PyCharm IDE files
+.idea/
+*.iml
+
+# System files
+.DS_Store
+Thumbs.db
